+++ example/lib/screens/forms/time_picker.dart
@@ -3,11 +3,9 @@ import 'package:example/widgets/page.dart';
 import 'package:fluent_ui/fluent_ui.dart';
 
 class TimePickerPage extends ScrollablePage {
-  PageState state = {
-    'simple_time': DateTime.now(),
-    'arrival_time': DateTime.now(),
-    '24_time': DateTime.now(),
-  };
+  DateTime? simpleTime;
+  DateTime? arrivalTime;
+  DateTime? hhTime;
 
   @override
   Widget buildHeader(BuildContext context) {
@@ -24,12 +22,9 @@ class TimePickerPage extends ScrollablePage {
       CardHighlight(
         child: Align(
           alignment: Alignment.centerLeft,
-          child: SizedBox(
-            width: 240.0,
-            child: TimePicker(
-              selected: state['simple_time'],
-              onChanged: (time) => setState(() => state['simple_tile'] = time),
-            ),
+          child: TimePicker(
+            selected: simpleTime,
+            onChanged: (time) => setState(() => simpleTime = time),
           ),
         ),
         codeSnippet: '''DateTime? selected;
@@ -47,14 +42,11 @@ TimePicker(
       CardHighlight(
         child: Align(
           alignment: Alignment.centerLeft,
-          child: SizedBox(
-            width: 240.0,
-            child: TimePicker(
-              header: 'Arrival time',
-              selected: state['arrival_time'],
-              onChanged: (time) => setState(() => state['arrival_time'] = time),
-              minuteIncrement: 15,
-            ),
+          child: TimePicker(
+            header: 'Arrival time',
+            selected: arrivalTime,
+            onChanged: (time) => setState(() => arrivalTime = time),
+            minuteIncrement: 15,
           ),
         ),
         codeSnippet: '''DateTime? selected;
@@ -74,14 +66,11 @@ TimePicker(
       CardHighlight(
         child: Align(
           alignment: Alignment.centerLeft,
-          child: SizedBox(
-            width: 240.0,
-            child: TimePicker(
-              header: '24 hour clock',
-              selected: state['24_time'],
-              onChanged: (v) => setState(() => state['24_time'] = v),
-              hourFormat: HourFormat.HH,
-            ),
+          child: TimePicker(
+            header: '24 hour clock',
+            selected: hhTime,
+            onChanged: (v) => setState(() => hhTime = v),
+            hourFormat: HourFormat.HH,
           ),
         ),
         codeSnippet: '''DateTime? selected;