+++ CHANGELOG.md
@@ -30,6 +30,24 @@ Date format: DD/MM/YYYY
 - Updated `TextBox`:
   - `TextBox` colors were updated to match the Win 11 design.
   - Fluent Text Selection Control now make use of `Acrylic`. Its items were also updated
+- Updated pickers ([#406](https://github.com/bdlukaa/fluent_ui/pull/406)):
+  - If `selected` is null, a placeholder text is shown ([#306](https://github.com/bdlukaa/fluent_ui/issues/306))
+  - Added new localization messages: `hour`, `minute`, `AM`, `PM`, `month`, `day`and `year`.
+  - **BREAKING** Removed `.hourPlaceholder`, `.minutePlaceholder`, `.amText`, `.pmText` from `TimePicker`. It was replaced, respectivelly, by the `hour`, `minute`, `AM`, `PM` localization messages
+  - On `DatePicker`, it's now possible to change the order of the fields:
+  ```dart
+  DatePicker(
+    ...,
+    fieldOrder: [
+      DatePickerField.day,
+      DatePickerField.month,
+      DatePickerField.year,
+    ],
+  )
+  ```
+
+  The fields are ordered based on the current locale by default
+  - On `DatePicker`, the day and year fields are now formatted based on the current locale (`getDateOrderFromLocale`)
 
 ## [4.0.0-pre.0] - [07/06/2022]